Sweeping !14600 from 21.3 to 21.9.
Nsw edm fixes

See merge request atlas/athena!14881

Former-commit-id: d978a0a0e8386ee69860a7c7c6382147e942e556

+++ MuonSpectrometer/MuonRDO/MuonRDO/STGC_RawData.h
@@ -19,9 +19,10 @@ class STGC_RawData {
  // offline identifier
    const Identifier m_id;
 
-   // readout data ( not sure whether time and charge should be float )
+   // readout data
+   // keeping time as a float for now, but it should also become an int
    float m_time;
-   float m_charge;
+   uint16_t m_charge;
    uint16_t m_bcTag;
    
    bool m_isDead;
@@ -31,7 +32,7 @@ class STGC_RawData {
    STGC_RawData (const Identifier id);
    STGC_RawData () {} //!< TODO remove this. Currently have problems with convertor if I do though.
 
-   STGC_RawData (const Identifier id, const uint16_t bcTag, const float time, const float charge, 
+   STGC_RawData (const Identifier id, const uint16_t bcTag, const float time, const uint16_t charge, 
 		 const bool isDead );
  
    virtual ~STGC_RawData() {}
@@ -40,7 +41,7 @@ class STGC_RawData {
 
    // access functions
    float time() const {return m_time;}
-   float charge() const {return m_charge;}
+   uint16_t charge() const {return m_charge;}
    uint16_t bcTag() const {return m_bcTag;}
 
    bool isDead() const {return m_isDead;}
